Fix #3339, Fix #3340, Fix #3341, Fix #3342: FNV ESM leaf/width corrections + ROADMAP corpus refresh
Four LOW-severity findings from AUDIT_FNV_2026-08-26 (dimensions 4 and 5).

#3339 — FactionRelation::combat_reaction was u8 while the XNAM parser read a
full u32 and cast with `as u8`. #482 widened the read specifically to survive
"a mod that extends the enum past 255", but the cast threw those bits away
immediately, leaving the wider read as nothing but cursor alignment. Store the
field at its on-disk u32 width. No live data changes: all 1,314 vanilla FNV
XNAM sub-records are 12 bytes with reaction values {0,1,2,3}. The existing
#482 test uses reaction=2, which fits either width and so cannot detect this;
the new test uses 0x00010002, whose low byte is also 2.

#3340 — expand_leveled_inner ran its `depth >= LVLI_MAX_DEPTH` guard before
the `index.items` push, so a terminal base item reached at exactly the cap was
discarded rather than collected. Pushing a leaf costs no further recursion, so
reorder the item check above the guard. This cannot weaken the cycle guard:
cycles run through `leveled_items`, never through `items`. The constant's doc
comment also claimed vanilla nesting "tops out around 3-4 levels" with
"comfortable headroom" at 8; a full walk of FalloutNV.esm (2,738 lists, 13,319
LVLO entries) measures a maximum depth of 7 — one level of margin, not
comfortable. Replace the estimate with the measured histogram and name the two
depth-7 vendor-chest roots.

#3341 — three FNV-unique record types (CCRD 270, CMNY 97, IMOD 69 = 436 of
13,319 LVLO leaves) dispatch into their own index maps and so fall through
expand_leveled_form_id's silent-skip branch. Dropping them is correct for the
equip use case — caravan cards, caravan money and weapon mods are not
wearable — so behaviour is unchanged. Add non_item_leaf_kind() to name the
record class at debug! when a leaf is indexed but not as an item, making the
boundary visible to a future container/loot consumer, for which a silent 3.3%
loss would be wrong. Genuinely unknown form IDs stay silent; the caller
already logs the originating outfit/NPC.

#3342 — the ROADMAP compat matrix still cited the pre-#3041 single-archive
corpus. cd9a5ef2 moved parse_rate_fallout_nv off game.mesh_archive() onto
open_all_mesh_archives; the FO3 and Skyrim SE gates moved with it. Re-measured
all three this session: FNV 20,746 NIFs across 11 archives (was cited as
14,881), FO3 17,172 across 6 (was 10,989), Skyrim SE 32,709 across 2 (was
18,862). All 100% clean, 0 truncated, 0 failed.

Full workspace suite green.

+++ crates/plugin/src/esm/records/actor/mod.rs
@@ -623,7 +623,16 @@ pub struct FactionRelation {
     /// Modifier (-100..100, larger means more friendly).
     pub modifier: i32,
     /// Combat reaction (0=neutral, 1=enemy, 2=ally, 3=friend).
-    pub combat_reaction: u8,
+    ///
+    /// Stored at the on-disk width (`u32` per UESP), not narrowed to the
+    /// vanilla 0..=3 range. Pre-#3339 this was `u8` while the parser read
+    /// a full `u32` and cast — so the `as u8` truncated back to exactly the
+    /// eight bits the pre-#482 `sub.data[8]` read had, defeating the stated
+    /// purpose of the wider read. Vanilla FNV never exceeds 3 (all 1,314
+    /// `XNAM` sub-records are 12 bytes; values `{0: 179, 1: 264, 2: 472,
+    /// 3: 399}`), so no live data changes — this only makes the field
+    /// honour what the parser actually decodes.
+    pub combat_reaction: u32,
 }
 
 #[derive(Debug, Clone)]
@@ -1611,13 +1620,16 @@ pub fn parse_fact(form_id: u32, subs: &[SubRecord], remap: &Option<FormIdRemap>)
             // The reaction field is a full 4-byte u32 per UESP; pre-#482 the
             // parser read only the low byte via `sub.data[8]`, which happened
             // to be correct for vanilla values 0..=3 but would silently
-            // truncate any future mod that extends the enum past 255.
+            // truncate any future mod that extends the enum past 255. The
+            // decoded value is stored at its full width (#3339) — the old
+            // `as u8` cast here threw away the extra 24 bits immediately,
+            // leaving the wider read purely a cursor-alignment step.
             b"XNAM" if sub.data.len() >= 8 => {
                 let mut r = SubReader::new(&sub.data);
                 let other = r.u32_or_default();
                 let modifier = r.i32_or_default();
                 let combat = if sub.data.len() >= 12 {
-                    r.u32_or_default() as u8
+                    r.u32_or_default()
                 } else {
                     0
                 };
